requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Google Ads API
google-ads==23.0.0
google-auth-oauthlib==1.2.0

# Data Processing
#pandas==2.1.4
#numpy==1.26.2

# CLI & Visualization
rich==13.7.0
click==8.1.7

# Web Dashboard
flask==3.0.0
flask-cors==4.0.0

# AI/ML (for future predictive features)
anthropic==0.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
